requirements.txt
python-telegram-bot==20.7
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
